vold: correct counting of partitions when NPARTS is not available.
This only effects devices using older kernels (before 2.6.32) which don't implement NPARTS/PARTN uevents.

If an sdcard has multiple paritions and NPARTS/PARTN are not implemented in the kernel vold will only attempt to
mount the last partition on the card, rather than attempt to mount each until it finds a FAT partition.
This is a problem as when there are multiple partitions present the last is usually not FAT.
With this patch vold correctly attempts to mount each partition.

Change-Id: If6fc4f5bbca1839f7a8d97b22bcb555e0e26797d